requirements.txt
astroid==2.4.2
boto==2.49.0
boto3==1.14.23
botocore==1.17.23
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
cryptography==2.9.2
docutils==0.15.2
ecdsa==0.15
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jwcrypto==0.7
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.5
pyasn1==0.4.8
pycparser==2.20
pyeviews==1.0.1
PyJWT==1.7.1
pylint==2.5.3
PyMySQL==0.9.3
python-dateutil==2.8.1
python-jose==3.1.0
pytz==2020.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1